api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
onnxruntime>=1.20.0
torch>=2.8.0
torchvision>=0.23.0
supabase>=1.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.4.0
matplotlib>=3.7.0
scipy>=1.10.0